Extract DecoderBuilderConfig into push_decoder.rs
Move the configuration struct and its builder construction out of
opener.rs into a new `push_decoder` module. The helper becomes a
`build(prepared_access_plan, metadata)` method on the config, which
reads more naturally at the call site.

No behavior change.
